requirements.txt
asgiref==3.4.1
Django==3.2
django-jalali-date==1.0.1
jdatetime==3.8.2
Pillow==8.4.0
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.1.1
